fix(collab): extract classification/material/ifcClass helpers out of from-ifcx.ts
Rebasing #3569 onto a moved main pushed from-ifcx.ts to 404 lines, past
check-module-size.mjs's 400-line new-module threshold. Split
setClassifications/setMaterials/readIfcClass into a sibling module
(packages/collab/src/snapshot/overlay-entity-attrs.ts). No behavior change.
